ml/requirements.txt
scikit-learn
pandas
psycopg2-binary
fastapi
gunicorn
uvicorn
pydantic